Revert "doc: evachargese: add documentation"
This reverts commit 36486ea68d1f836005b8d68b2aa82c0948b1b525.

Reason for revert is that read-the-docs.org has currently problems when
cloning the stib repository and thus it's not possible to generate the
documentation. As a workaround the documentation is moved into
a stand-alone repository: https://github.com/i2se/evacharge-se-bsp

I'm removing the docs here entirely, so that we do not need to
maintain two copies which would lead to confusion etc.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>